requirements.txt
asgiref==3.2.10
Django==3.2.16
django-environ==0.4.5
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1

djangorestframework~=3.14.0
environ~=1.0